Tighten up contextsave, contextload, and taskstart
This ensures we always return 0 or 1 from contextsave.

This turns out to be a bit tricky, becuase contextload needs to return
to taskstart when creating a new task, which has its own function
signature.

So, for now, taskstart includes a hole in its arguments where
contextload can safely write without interfering.

In the future, it may make sense to instead make taskstart match
`func (uintptr) int` in its signature.